tidy up state module
